+++ unicore_gitmodels/models.py
@@ -1,6 +1,8 @@
 import re
 import unicodedata
 from gitmodel import fields, models
+from unicore_gitmodels.fields import ListField
+
 
 RE_NUMERICAL_SUFFIX = re.compile(r'^[\w-]*-(\d+)+$')
 
@@ -44,6 +46,7 @@ class GitCategoryModel(SlugifyMixin, FilterMixin, models.GitModel):
     title = fields.CharField(required=True)
     subtitle = fields.CharField(required=False)
     language = fields.CharField(required=False)
+    featured_in_navbar = fields.BooleanField(default=False)
     source = fields.RelatedField('GitCategoryModel', required=False)
 
     def __unicode__(self):
@@ -78,6 +81,7 @@ class GitCategoryModel(SlugifyMixin, FilterMixin, models.GitModel):
             'title': self.title,
             'subtitle': self.subtitle,
             'language': self.language,
+            'featured_in_navbar': self.featured_in_navbar,
             'source': source,
         }
 
@@ -96,6 +100,7 @@ class GitPageModel(SlugifyMixin, FilterMixin, models.GitModel):
     featured_in_category = fields.BooleanField(default=False)
     language = fields.CharField(required=False)
     source = fields.RelatedField('GitPageModel', required=False)
+    linked_pages = ListField(default=[], required=False)
 
     def __unicode__(self):
         return self.title
@@ -122,6 +127,13 @@ class GitPageModel(SlugifyMixin, FilterMixin, models.GitModel):
             'published': self.published,
             'primary_category': primary_category,
             'source': source,
+            'language': self.language,
             'featured': self.featured,
             'featured_in_category': self.featured_in_category,
+            'linked_pages': self.linked_pages,
         }
+
+    def get_linked_pages(self):
+        if self.linked_pages is None:
+            return []
+        return [self.__class__.get(uuid) for uuid in self.linked_pages]