+++ unicore_gitmodels/tests/test_models.py
@@ -87,10 +87,11 @@ class ModelsTestCase(unittest.TestCase):
         self.assertEquals(p2.source.uuid, p.uuid)
 
     def mk_category(self, title, subtitle='subtitle',
-                    language=None, source=None):
+                    language=None, source=None, featured=False):
         models = self.get_repo_models()
         category = models.GitCategoryModel(
-            title=title, subtitle=subtitle, language=language, source=source)
+            title=title, subtitle=subtitle, language=language, source=source,
+            featured_in_navbar=featured)
         category.save(True)
         return category
 
@@ -99,7 +100,7 @@ class ModelsTestCase(unittest.TestCase):
                 created_at=None, modified_at=None,
                 published=True, primary_category=None,
                 featured_in_category=None, language='eng-US',
-                source=None, featured=None):
+                source=None, featured=None, linked_pages=None):
         models = self.get_repo_models()
         now = modified_at or datetime.now()
         then = created_at or (now - timedelta(days=1))
@@ -109,19 +110,22 @@ class ModelsTestCase(unittest.TestCase):
             content=content, created_at=then, modified_at=now,
             published=published, primary_category=primary_category,
             featured_in_category=featured_in_category,
-            language=language, source=source, featured=featured)
+            language=language, source=source, featured=featured,
+            linked_pages=linked_pages)
         page.save(True)
         return page
 
     def test_category_to_dict(self):
         category1 = self.mk_category('category1')
-        category2 = self.mk_category('category2', source=category1)
+        category2 = self.mk_category(
+            'category2', source=category1, featured=True)
         self.assertEquals(category2.to_dict(), {
             'uuid': category2.uuid,
             'title': u'category2',
             'subtitle': 'subtitle',
             'slug': u'category2',
             'language': '',
+            'featured_in_navbar': True,
             'source': category1.to_dict(),
         })
 
@@ -148,4 +152,18 @@ class ModelsTestCase(unittest.TestCase):
             'source': page1.to_dict(),
             'featured_in_category': True,
             'featured': True,
+            'language': 'eng-US',
+            'linked_pages': None,
         })
+
+    def test_linked_pages(self):
+        page1 = self.mk_page('page1')
+        page2 = self.mk_page('page2')
+        page3 = self.mk_page('page3', linked_pages=[page1.uuid, page2.uuid])
+
+        data = page3.to_dict()
+        self.assertEquals(data['linked_pages'], [page1.uuid, page2.uuid])
+
+        linked_page1, linked_page2 = page3.get_linked_pages()
+        self.assertEqual(linked_page1.to_dict(), page1.to_dict())
+        self.assertEqual(linked_page2.to_dict(), page2.to_dict())